refactor: share generateUpdateMany between the pg and sqlite builders
The `update<Table>Many` generator was duplicated between the pg and sqlite
builders (179 and 220 lines). The two are identical from the argument map
through the per-entry remap and validation, the nested-write split, the
write hooks, relation enrichment and the per-entry output slots — roughly
150 lines. They differ only in how the batch executes inside its
transaction.

`src/util/builders/update-many.ts` holds the shared resolver, and that one
difference is now the `UpdateManyBatchRunner` hook:

- `runUpdateManyBatch` (the default) is what pg used: one UPDATE per entry,
  awaited in input order, inside one `executor.transaction`.
- `runSqliteUpdateManyBatch` stays in `sqlite.ts`, because a synchronous
  driver (better-sqlite3) commits the moment its transaction callback
  returns. It cannot hand the driver an async callback, so it probes the
  driver kind from the first statement's result and runs the entries
  synchronously when there is nothing to await. That is genuinely
  sqlite-specific, not drift.

pg's generator is now one line; sqlite keeps its runner and nothing else.

No behavior change. Verified with `npm run typecheck` (both projects),
biome, the 70-file pglite/sqlite suite (1044 tests) and the Docker pg/mysql
suite (180 tests).
